requirements.txt
httpx>=0.27.0
chromadb>=0.5.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
streamlit>=1.35.0
tenacity>=8.3.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
pytest>=8.2.0
eval_type_backport>=0.2.0
